+++ services/preview/compute/mgmt/2020-10-01-preview/compute/CHANGELOG.md
@@ -1,2 +1,9 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Signature Changes
+
+#### Struct Fields
+
+1. OrchestrationServiceStateInput.ServiceName changed type from OrchestrationServiceNames to *string